soap/requirements.txt
Flask
requests
bs4
Flask-Spyne
